chore: Move `set_suggestions` fuction to records.py API module
